delete HTMLMixin and have write_html() available on FPDF class
