+++ fpdf/fpdf.py
@@ -23,6 +23,7 @@ import re
 import sys
 import warnings
 import zlib
+import html
 from collections import OrderedDict, defaultdict
 from collections.abc import Sequence
 from contextlib import contextmanager
@@ -35,6 +36,7 @@ from typing import Callable, List, NamedTuple, Optional, Tuple, Union
 from fontTools import ttLib
 from fontTools import subset as ftsubset
 from io import BytesIO
+from fpdf.html import HTML2FPDF
 
 try:
     from PIL.Image import Image
@@ -504,6 +506,15 @@ class FPDF(GraphicsStateMixin):
         self.record_text_quad_points = False
         # page number -> array of 8 × n numbers:
         self.text_quad_points = defaultdict(list)
+
+    def write_html(self, text, *args, **kwargs):
+        """Parse HTML and convert it to PDF"""
+        kwargs2 = vars(self)
+        # Method arguments must override class & instance attributes:
+        kwargs2.update(kwargs)
+        h2p = HTML2FPDF(self, *args, **kwargs2)
+        text = html.unescape(text)  # To deal with HTML entities
+        h2p.feed(text)
 
     def _add_quad_points(self, x, y, w, h):
         self.text_quad_points[self.page].extend(
